bugfixing + refactor
