--- src/api/const.ts
@@ -1,6 +1,0 @@
-export const ApiRoutes = {
-  Offers: '/offers',
-};
-
-export const API_TIMEOUT = 5000;
-export const API_URL = 'https://16.design.htmlacademy.pro/six-cities';
