--- src/api/types.ts
@@ -1,8 +1,0 @@
-import {AppDispatch, RootState} from '../store';
-import {AxiosInstance} from 'axios';
-
-export type ThunkOptions = {
-  dispatch: AppDispatch;
-  state: RootState;
-  extra: AxiosInstance;
-}
