+++ src/components/location-list/location-list.tsx
@@ -1,6 +1,6 @@
 import { useEffect } from 'react';
 import { useSearchParams } from 'react-router-dom';
-import { LOCATIONS, DEFAULT_CITY } from '../../const';
+import { LOCATIONS, DEFAULT_CITY, URLParams } from '../../const';
 import { CityName } from '../../types';
 import { changeCity, selectActiveCity } from '../../store/slices/city';
 import { useAppDispatch, useAppSelector } from '../../store/hooks';
@@ -14,7 +14,7 @@ export default function LocationList(): JSX.Element {
   const locations: CityName[] = Object.values(LOCATIONS) as CityName[];
 
   useEffect(() => {
-    const cityFromURL = searchParams.get('city') as CityName;
+    const cityFromURL = searchParams.get(URLParams.City) as CityName;
 
     if (cityFromURL && locations.includes(cityFromURL)) {
       dispatch(changeCity(cityFromURL));
