+++ src/const.ts
@@ -119,3 +119,13 @@ export enum Slices {
   City = 'city',
 }
 
+export enum URLParams {
+  City = 'city',
+}
+
+export const ApiRoutes = {
+  Offers: '/offers',
+};
+
+export const API_TIMEOUT = 5000;
+export const API_URL = 'https://16.design.htmlacademy.pro/six-cities';
