+++ src/store/slices/city.ts
@@ -1,7 +1,7 @@
 import { createSlice, PayloadAction } from '@reduxjs/toolkit';
 import { CityName } from '../../types';
 import { RootState } from '../index';
-import { DEFAULT_CITY, LOCATIONS } from '../../const';
+import { DEFAULT_CITY, LOCATIONS, URLParams } from '../../const';
 import { Slices } from '../../const';
 
 export interface State {
@@ -10,7 +10,7 @@ export interface State {
 
 const getInitialCity = (): CityName => {
   const params = new URLSearchParams(window.location.search);
-  const cityFromURL = params.get('city') as CityName;
+  const cityFromURL = params.get(URLParams.City) as CityName;
 
   if (cityFromURL && Object.values(LOCATIONS).includes(cityFromURL)) {
     return cityFromURL;
