+++ src/store/slices/loading.ts
@@ -16,7 +16,7 @@ const loadingSlice = createSlice({
   name: Slices.Loading,
   initialState,
   reducers: {
-    getLoading(state, action: PayloadAction<boolean>) {
+    setLoading(state, action: PayloadAction<boolean>) {
       state.loading = action.payload;
     },
     setError(state, action: PayloadAction<boolean>) {
@@ -25,7 +25,7 @@ const loadingSlice = createSlice({
   },
 });
 
-export const { getLoading, setError } = loadingSlice.actions;
+export const { setLoading, setError } = loadingSlice.actions;
 
 export const selectLoading = (state: RootState): boolean => state.loading.loading;
 export const selectError = (state: RootState): boolean => state.loading.error;
