+++ src/types.ts
@@ -55,3 +55,12 @@ export type sortOptionNamesType = {
 export type SortOptionKey = keyof typeof sortOptionNames;
 
 export type SortOptionName = typeof sortOptionNames[SortOptionKey];
+
+import {AppDispatch, RootState} from './store';
+import {AxiosInstance} from 'axios';
+
+export type ThunkOptions = {
+  dispatch: AppDispatch;
+  state: RootState;
+  extra: AxiosInstance;
+}
